Moved InitType enum from field_utils.hpp to field_tracking.hpp
