+++ components/scream/src/share/field/field_tracking.hpp
@@ -1,7 +1,6 @@
 #ifndef SCREAM_FIELD_TRACKING_HPP
 #define SCREAM_FIELD_TRACKING_HPP
 
-#include "share/field/field_utils.hpp"
 #include "share/scream_types.hpp"
 
 #include "share/util/scream_time_stamp.hpp"
@@ -14,6 +13,29 @@
 #include <set>
 
 namespace scream {
+
+// How a field has to be init-ed
+// Note: internally, Value still creates a FieldInitializer object, but can be done
+//       behind the scenes by the infrastructure
+enum class InitType {
+  // NotNeeded,    // No initialization is needed for this field
+  Value,        // Should be inited to a specific value
+  Initializer,  // A FieldInitializer object should take care of this
+  None          // No initialization is needed/expected
+};
+
+inline std::string e2str (const InitType e) {
+  std::string s;
+  switch (e) {
+    // case InitType::NotNeeded:   s = "NotNeeded";    break;
+    case InitType::None:        s = "None";         break;
+    case InitType::Value:       s = "Value";        break;
+    case InitType::Initializer: s = "Initializer";  break;
+    default: s = "INVALID";
+  }
+
+  return s;
+}
 
 // Forward declarations
 class AtmosphereProcess;
